Finish: simple allocator

+++ kernel8/include/allocator.h
@@ -1,1 +1,1 @@
-char *simple_alloc(unsigned int nbytes);
+char *simple_alloc(unsigned long nbytes);
